Change directory tree and remove error.go
