--- pkg/recoverable/error.go
@@ -1,9 +1,0 @@
-package recoverable
-
-import (
-	"fmt"
-)
-
-func toError(r interface{}) error {
-	return fmt.Errorf("%v", r)
-}
