+++ recoverable/runner.go
@@ -1,13 +1,13 @@
 package recoverable
 
 // Runner will run f function, and rerun f if panic occurs in f
-func Runner(f func(), panicHandler func(e error)) {
+func Runner(f func(), panicHandler func(r interface{})) {
 	running := true
 	for running {
 		func() {
 			defer func() {
 				if r := recover(); r != nil {
-					panicHandler(toError(r))
+					panicHandler(r)
 				}
 			}()
 			f()
