+++ recoverable/waitgroup.go
@@ -6,13 +6,13 @@ import (
 
 type WaitGroup struct {
 	wg   sync.WaitGroup
-	ch   chan error
+	ch   chan interface{}
 	once sync.Once
 }
 
 func (wg *WaitGroup) Add(n int) {
 	wg.once.Do(func() {
-		wg.ch = make(chan error)
+		wg.ch = make(chan interface{})
 	})
 	wg.wg.Add(n)
 }
@@ -20,7 +20,7 @@ func (wg *WaitGroup) Add(n int) {
 func (wg *WaitGroup) Done() {
 	defer wg.wg.Done()
 	if r := recover(); r != nil {
-		wg.ch <- toError(r)
+		wg.ch <- r
 	}
 }
 
@@ -29,11 +29,11 @@ func (wg *WaitGroup) Wait() {
 		wg.wg.Wait()
 		close(wg.ch)
 	}()
-	var errs []error
-	for err := range wg.ch {
-		errs = append(errs, err)
+	var rs []interface{}
+	for r := range wg.ch {
+		rs = append(rs, r)
 	}
-	if len(errs) > 0 {
-		panic(errs)
+	if len(rs) > 0 {
+		panic(rs)
 	}
 }
